Limma Hyperparameters Screening

This function screens through various combinations of hyperparameters for limma analysis, including designs, modes, and degrees of freedom. It validates inputs, generates results for all combinations, and plots the outcomes. Finally, it may also be involved in generating an HTML report as part of a larger analysis workflow.

limma_hyperparams_screen(
  datas,
  datas_descr,
  metas,
  designs,
  modes,
  condition,
  spline_configs,
  feature_names,
  report_dir = here::here(),
  adj_pthresh = c(0.05),
  padjust_method = "BH"
)

Arguments

datas

A list of data frames containing the datasets to be analyzed.

metas

A list of data frames containing metadata for each dataset in `datas`.

designs

A character vector of design formulas for the limma analysis.

modes

A character vector indicating the mode of analysis for each design.

feature_names

A character vector of feature names to be analyzed.

padjust_method

A character string specifying the method for p-value adjustment.

factors

A character vector of factors to be considered in the analysis.

DoFs

An integer vector of degrees of freedom to be used for each factor in `factors`.

pthresholds

A numeric vector of p-value thresholds for significance determination.

Value

Returns a list of plots generated from the limma analysis results. Each element in the list corresponds to a different combination of hyperparameters.

Examples

if (FALSE) {
  results <- limma_hyperparams_screen(datas, metas, designs, modes, factors,
                                      DoFs = c(2L), feature_names, 
                                      pthresholds = c(0.05),
                                      padjust_method = "BH")
}